erreur formul

+++ appP5/front/JS/panier.js
@@ -66,16 +66,7 @@ let mailVal = mail.value;
 let mailValue;
 
 //fonction de vérificatiçn de la validité des champs du formulaire
-function verifChamps(regex, value) {
-  let alert = document.getElementById('message');
-  let verif = regex.test(value);
-  if (verif == false) {
-    alert.innerHTML = "champs non conforme !!";
-    console.log("champs non comforme");
-  } else {
-    alert.innerHTML = " ";
-  };
-}
+
 
 //création de l'objet contact à envoyer au serveur pour la commande
 let contact;
@@ -83,26 +74,54 @@ let contact;
 //Application des regex sur les input du formulaire
 function verifFormulaire() {
 
-  let regexNom = /[a-z]{3}/;
-  let regexMail = /[a-z]+@+/;
-  
-var list = document.getElementsByClassName("formelt");
-for (let item of list) {
-    console.log(item.id);
-  item.addEventListener('blur', function (e) {
-    let preValue = e.target.value;
-    console.log("preValue " + preValue);
-    verifChamps(regexNom, preValue);
-    });
-  }
-  
-    mail.addEventListener("focus", function (e) {
+
+  function verifChamps(regex, value) {
+    let alert = document.getElementById('message');
+    let verif = regex.test(value);
+    if (verif == false) {
+      alert.innerHTML = "champs non conforme !!";
+      console.log("champs non comforme");
+    } else {
+      alert.innerHTML = " ";
+    };
+  };
+
+  let regexNom = /[Aa-zZ]{3}/;
+  let regexMail = /.+@.+\..+/;
+
+  prenom.addEventListener("blur", function (e) {
+    prenomValue = e.target.value;
+    console.log("prenomValue " + prenomValue);
+    verifChamps(regexNom, prenomValue);
+  });
+
+  nom.addEventListener("blur", function (e) {
+    nomValue = e.target.value;
+    console.log("nomValue : " + nomValue);
+    verifChamps(regexNom, nomValue);
+  });
+
+  adresse.addEventListener("blur", function (e) {
+    adresseValue = e.target.value;
+    console.log(" adresseValue : " + adresseValue);
+    verifChamps(regexNom, adresseValue);
+  });
+
+  ville.addEventListener("blur", function (e) {
+    villeValue = e.target.value;
+    console.log(" villeValue : " + villeValue);
+    verifChamps(regexNom, villeValue);
+  });
+
+  mail.addEventListener("blur", function (e) {
     mailValue = e.target.value;
+    console.log("mailValue : " + mailValue);
     verifChamps(regexMail, mailValue);
-    document.getElementById('go').style.display = "block";
   });
   
+  
 }
+
 
 verifFormulaire();
 
@@ -120,21 +139,32 @@ function contacter() {
   mailValue = mail.value;
   console.log('mail value contact : ' + mailValue);
 
-  contact = {
-    firstName: preValue,
-    lastName: nomValue,
-    address: adrValue,
-    city: villeValue,
-    email: mailValue
+  if (!preValue || !nomValue || !adrValue || !villeValue || !mailValue) {
+    let reset = document.getElementById('reset');
+  
+    reset.innerHTML = '<p class="errorCommande"> UNE ERREUR INCONNUE C\'EST PRODUITE.<br>Veulliez nous excusez pour la gène occasionnée.<br>Nous vous redirigeons sur votre boutique.<br> N\'hesitez pas à renouveller votre commande. </p>';
+      localStorage.clear();
+      setTimeout(function(){window.history.back()}, 5000) ;
+  } else {
+
+    contact = {
+      firstName: preValue,
+      lastName: nomValue,
+      address: adrValue,
+      city: villeValue,
+      email: mailValue
+    }
+    return contact;
   }
-  return contact;
 };
 
 console.log(" contact: " + contact);
 
 //requete finale de commande contenant les informations de contact et les IDs produit
 function achat() {
+
   contacter();
+
   let objt = {
     contact,
     products
@@ -142,13 +172,25 @@ function achat() {
   console.log("objt : " + objt);
   let buy = JSON.stringify(objt);
   console.log(buy);
+
   let request = new XMLHttpRequest();
   request.onreadystatechange = function () {
     if (this.readyState == XMLHttpRequest.DONE) {
       let confirmation = JSON.parse(this.responseText);
       console.log(confirmation);
+//     let conretour = JSON.stringify(confirmation.contact)
+//     let formretour = JSON.parse(conretour);
+//      console.log(formretour);
+//      for (let coor of formretour){
+//          console.log(coor[i]);
+//      }
+//     
+//      
+      
+      
+      
       let clear = document.getElementById('contact');
-      clear.innerHTML="Votre commande sera traitée dans les plus bref délais. ";
+      clear.innerHTML = "Votre commande sera traitée dans les plus bref délais. ";
       document.getElementById('ticket').innerHTML += "<p class=\"ticket\"> Votre commande à bien été prise en compte sous le numéro :  <br>" + confirmation.orderId + "<br>Merci de votre visite sur ORINOCO</p>";
     };
   }
@@ -160,8 +202,13 @@ function achat() {
 //bouton "acheter" pour envoyer le formulire de commande, vide le localstorage.
 let send = document.getElementById('go');
 send.addEventListener('click', function () {
-  console.log('ok');
-  achat();
-  localStorage.clear();
+  let message = document.getElementById('message');
+  if (message = " " ) {
+    achat();
+    localStorage.clear();
+    console.log('ok');
+  } else {
+    alert("Le formulaire est mal renseigné");
+  }
 
 });
